Refactored handling of Framebuffer read/update in FPGA listener
Made to be able to move function into nChip16 emulator in the future.
Also added a setting in nChip16 to enable/disable Listener. Either you use the emulator to output the picture by running the emulator or you leave the emulator in Pause mode and enable updating the UI using Xilinx ISE FPGA Simulation software